requirements.txt
torch>=2.0.0
numpy>=1.21.0
python-chess>=1.999
